[Refactor] Move example-only helpers out of tilelang package
