+++ examples/dequantize_gemm/example_dequant_gemm_bf16_mxfp4_hopper.py
@@ -154,7 +154,7 @@ def matmul(
     B_dequantize_shared_shape = (block_N, block_K)
     assert K % (block_K * split) == 0
 
-    from tilelang.quantize import get_mxfp_intrin_group
+    from quantize import get_mxfp_intrin_group
 
     # fast_dequant_bf16_fp4_twiddling
     mxfp_intrin_info = get_mxfp_intrin_group(
